services/trades/pyproject.toml
[project]
name = "trades"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "pydantic-settings>=2.6.1",
    "quixstreams>=3.4.0",
    "websocket-client>=1.8.0",
]
